Make adding towers from the sidebar possible
Previously the pictures of the towers in the sidebar did not follow the mouse properly. Now they snap to the mouse. The player can now add a tower if the she clicks on a button in the sidebar. TODO: the towerpictures should be the same size as the buttons underneath. Otherwise only the picture is moved resulting in a nasty bug where you cannot get rid of the picture underneath the mouse.